lifecycle: split into its own package (#1626)
This is a small refactor to make the lifecycle module
a bit more manageable and separate some concerns
for the logic within it.
