+++ snapcraft/internal/lifecycle/__init__.py
@@ -1,0 +1,22 @@
+# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
+#
+# Copyright (C) 2017 Canonical Ltd
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+from ._clean import clean                 # noqa
+from ._containers import cleanbuild       # noqa
+from ._containers import containerbuild   # noqa
+from ._init import init                   # noqa
+from ._packer import pack                 # noqa
+from ._packer import snap                 # noqa
+from ._runner import execute              # noqa
